# docs/swagger.yaml
openapi: 3.0.0
info:
  title: MyZubster Gateway API
  description: API Gateway per MyZubster - MCP Server per AI Agents
  version: 1.0.0
  contact:
    name: MyZubster Team
    url: https://myzubster.com
    email: support@myzubster.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.myzubster.com
    description: Production Server
  - url: http://localhost:5002
    description: Development Server
...
paths:
...
  /api/tokens/balance/{userId}:
    get:
      summary: Saldo token
      description: Ottieni il saldo dei token per un utente
      tags:
        - Wallet
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: ID dell'utente
      responses:
        '200':
          description: Saldo token
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                  balances:
                    type: object
                    properties:
                      MYZ:
                        type: number
                      XMR:
                        type: number
                      BTC:
                        type: number
                      ETH:
                        type: number
                      ADA:
                        type: number
                  totalUSD:
                    type: number
                  totalSGD:
                    type: number
                  timestamp:
                    type: string
                    format: date-time
...